requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
PySide6>=6.6.0
faster-whisper>=1.0.0
sounddevice>=0.4.6
